+++ index.html
@@ -235,8 +235,5 @@
 
   <script src="vendor/pristine/pristine.min.js"></script>
   <script src="js/main.js" type="module"></script>
-  <script src="validation/validation.js" type="module"></script>
-  <script src="validation/rules.js" type="module"></script>
-  <script src="js/forms.js" type="module"></script>
 </body>
 </html>
