+++ js/validation/validation.js
@@ -1,6 +1,5 @@
-import { HASHTAG, MAX_COUNT_HASTAG} from './rules.js';
+import { HASHTAG} from './rules.js';
 
-//Валидируем уникальность хэштегов
 const validateUniqueTags = (tags) => {
   const lowerCaseTags = tags.map((tag) => tag.toLowerCase());
   return new Set(lowerCaseTags).size === lowerCaseTags.length;
@@ -8,8 +7,6 @@ const validateUniqueTags = (tags) => {
 
 export const validateLength = (maxLength) => (iter) => iter.length <= maxLength;
 
-
-//Валидируем остальные требования к тегу
 const isValidTag = (tag) => HASHTAG.test(tag);
 
 export const validateTags = (value) => {
@@ -20,5 +17,5 @@ export const validateTags = (value) => {
     .trim()
     .split(' ')
     .filter((tag) => tag.trim().length);
-  return validateLength(MAX_COUNT_HASTAG) && validateUniqueTags(tags) && tags.every(isValidTag);
+  return validateLength(tags) && validateUniqueTags(tags) && tags.every(isValidTag);
 };
